Refactor jukeboxService : first step (WIP). JukeboxService is now in front of both JukeboxJavaService and JukeboxLegacySubsonicService.
