+++ airsonic-main/src/main/java/org/airsonic/player/controller/RESTController.java
@@ -111,8 +111,6 @@ public class RESTController {
     @Autowired
     private JukeboxService jukeboxService;
     @Autowired
-    private JukeboxJavaService jukeboxJavaService;
-    @Autowired
     private AudioScrobblerService audioScrobblerService;
     @Autowired
     private PodcastService podcastService;
@@ -846,8 +844,6 @@ public class RESTController {
         }
 
         Player player = playerService.getPlayer(request, response);
-        boolean isJavaJukebox = player.getTechnology().equals(PlayerTechnology.JAVA_JUKEBOX);
-
 
         boolean returnPlaylist = false;
         String action = getRequiredStringParameter(request, "action");
@@ -874,11 +870,7 @@ public class RESTController {
             playQueueService.doShuffle(request, response);
         } else if ("setGain".equals(action)) {
             float gain = getRequiredFloatParameter(request, "gain");
-            if (isJavaJukebox) {
-                jukeboxJavaService.setGain(player,gain);
-            } else {
-                jukeboxService.setGain(gain);
-            }
+                jukeboxService.setGain(player,gain);
         } else if ("get".equals(action)) {
             returnPlaylist = true;
         } else if ("status".equals(action)) {
@@ -890,26 +882,15 @@ public class RESTController {
         String username = securityService.getCurrentUsername(request);
         PlayQueue playQueue = player.getPlayQueue();
 
-        boolean controlsJukebox=false;
-        if (isJavaJukebox) {
-            controlsJukebox = true;
-        }  else {
-            Player jukeboxPlayer = jukeboxService.getPlayer();
-            controlsJukebox = jukeboxPlayer != null && jukeboxPlayer.getId().equals(player.getId());
-        }
-
+        // this variable is only needed for the JukeboxLegacySubsonicService. To be removed.
+        boolean controlsJukebox = jukeboxService.canControl(player);
 
         int currentIndex = controlsJukebox && !playQueue.isEmpty() ? playQueue.getIndex() : -1;
         boolean playing = controlsJukebox && !playQueue.isEmpty() && playQueue.getStatus() == PlayQueue.Status.PLAYING;
         float gain;
         int position;
-        if (isJavaJukebox) {
-            gain = jukeboxJavaService.getGain(player);
-            position = controlsJukebox && !playQueue.isEmpty() ? jukeboxJavaService.getPosition(player) : 0;
-        } else {
-            gain = jukeboxService.getGain();
-            position = controlsJukebox && !playQueue.isEmpty() ? jukeboxService.getPosition() : 0;
-        }
+        gain = jukeboxService.getGain(player);
+        position = controlsJukebox && !playQueue.isEmpty() ? jukeboxService.getPosition(player) : 0;
 
         Response res = createResponse();
         if (returnPlaylist) {
